Add lobby-specific pool page
The ref bot will link to this page when a lobby is created, so it can display the actual maps as seen by the bot. This should avoid any desyncs that may happen if a player clicks the pool link and a map's rating has been updated since the bot got the mappool. Leading to the player seeing a different map than what the bot is using.